Update README.md
This Point of Sales is integrated with an Inventory Management system. It allows a registered user access according to their job title, where a registered waiter is only allowed to order a meal and a manager is allowed full access to all functionalities of the system. The functionalities of the system include being able to place an order, check inventory, adding new users, updating stock and many more.